UI-7 move files around
