+++ src/Nav/SideNav/components/SideNavLink/index.scss
@@ -1,4 +1,4 @@
-@import '../../../style/colors';
+@import '../../style/colors';
 
 .navbar-nav {
   opacity: 0.6;
