+++ src/Nav/SideNav/index.scss
@@ -1,0 +1,3 @@
+#sidenav {
+  width: 200px;
+}
