+++ src/Nav/SideNav/index.ts
@@ -1,0 +1,6 @@
+import { SideNav, INavLink } from './SideNav'
+
+export {
+  SideNav,
+  INavLink,
+}
