+++ src/Nav/index.scss
@@ -1,3 +1,15 @@
 .layout {
   background: rgb(246, 247, 251);
+  min-height: 100vh;
+  flex: 1 1 auto;
+  display: flex;
+  flex-direction: column;
+}
+
+.main-content {
+  flex: 1 1 auto;
+}
+
+.content {
+  flex: 1 1 auto;
 }
