--- src/Navbar/index.scss
@@ -1,19 +1,0 @@
-.navbar-light {
-  background-color: white;
-}
-
-.navbar-brand {
-  color: white !important;
-}
-
-.doseme-logo {
-  width: 200px;
-}
-
-.modal-dialog {
-  max-width: 800px;
-}
-
-.cursor {
-  cursor: pointer;
-}
