--- src/Navbar/index.ts
@@ -1,6 +1,0 @@
-import { Navbar, INavLink } from './Navbar'
-
-export {
-  Navbar,
-  INavLink
-}
