--- src/SideNav/index.scss
@@ -1,3 +1,0 @@
-#sidenav {
-  width: 200px;
-}
