--- src/SideNav/index.ts
@@ -1,5 +1,0 @@
-import { SideNav } from './SideNav'
-
-export {
-  SideNav
-}
