+++ src/index.ts
@@ -1,16 +1,13 @@
 import { AdminLogo } from './AdminLogo'
-import { SideNav } from './SideNav'
-import { Navbar, INavLink } from './Navbar'
 import { Footer } from './Footer'
 import { Panel } from './Panel'
 import { SmartList } from './SmartList'
+import * as Nav from './Nav'
 
 export {
   AdminLogo,
   Footer,
-  Navbar,
   Panel,
-  SideNav,
-  INavLink,
+  Nav,
   SmartList,
 }
